remove deprecated component move mock data code clean up
