--- app/routes/screens/race/components/styles.js
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export default StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center"
-  },
-  text: {
-    color: "white",
-    fontSize: 45,
-    backgroundColor: "transparent"
-  }
-});
